requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.10.8
django-autofixture==0.12.1
django-haystack==2.6.1
django-markdown==0.8.4
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.4
pysolr==3.6.0
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
solr==0.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
